requirements.txt
Django~=4.1
django-oauth-toolkit~=2.1.0
requests~=2.28.1
djangorestframework~=3.13.1
python-dotenv~=0.20.0
django-cors-headers~=3.13.0
git+https://github.com/maurice-herwig/wofa
